+++ visualization/app/codeCharta/codeCharta.component.ts
@@ -2,17 +2,16 @@ import { UrlExtractor } from "./util/urlExtractor"
 import { IHttpService, ILocationService } from "angular"
 import "./codeCharta.component.scss"
 import { CodeChartaService } from "./codeCharta.service"
-import { ScenarioHelper } from "./util/scenarioHelper"
 import { DialogService } from "./ui/dialog/dialog.service"
-import { NameDataPair } from "./codeCharta.model"
+import { NameDataPair, SearchPanelMode } from "./codeCharta.model"
 import { InjectorService } from "./state/injector.service"
 import { StoreService } from "./state/store.service"
-import { setState } from "./state/store/state.actions"
 import { setAppSettings } from "./state/store/appSettings/appSettings.actions"
 import { setIsLoadingFile } from "./state/store/appSettings/isLoadingFile/isLoadingFile.actions"
 import * as codeCharta from "../../package.json"
 import { setDelta, setMultiple, setSingle } from "./state/store/files/files.actions"
 import { getCCFiles } from "./model/files/files.helper"
+import { setSearchPanelMode } from "./state/store/appSettings/searchPanelMode/searchPanelMode.actions"
 
 export class CodeChartaController {
 	private _viewModel: {
@@ -67,19 +66,15 @@ export class CodeChartaController {
 		])
 	}
 
+	public onClick() {
+		if (this.storeService.getState().appSettings.searchPanelMode !== SearchPanelMode.minimized) {
+			this.storeService.dispatch(setSearchPanelMode(SearchPanelMode.minimized))
+		}
+	}
+
 	private tryLoadingFiles(values: NameDataPair[]) {
 		this.storeService.dispatch(setAppSettings())
-
-		this.codeChartaService
-			.loadFiles(values)
-			.then(() => {
-				this.storeService.dispatch(setState(ScenarioHelper.getDefaultScenarioSetting()))
-			})
-			.catch(e => {
-				this.storeService.dispatch(setIsLoadingFile(false))
-				console.error(e)
-				this.printErrors(e)
-			})
+		this.codeChartaService.loadFiles(values)
 	}
 
 	private setRenderStateFromUrl() {
@@ -93,10 +88,6 @@ export class CodeChartaController {
 		} else {
 			this.storeService.dispatch(setSingle(files[0]))
 		}
-	}
-
-	private printErrors(errors: Object) {
-		this.dialogService.showErrorDialog(JSON.stringify(errors, null, "\t"))
 	}
 }