+++ visualization/app/codeCharta/ui/dialog/dialog.download.component.spec.ts
@@ -14,7 +14,7 @@ import { setMarkedPackages } from "../../state/store/fileSettings/markedPackages
 
 describe("DialogDownloadController", () => {
 	stubDate(new Date(Date.UTC(2018, 11, 14, 9, 39)))
-	const newDate: string = "2018-12-14_09-39"
+	const newDate = "2018-12-14_09-39"
 
 	let dialogDownloadController: DialogDownloadController
 	let $mdDialog
@@ -40,12 +40,8 @@ describe("DialogDownloadController", () => {
 	}
 
 	function withMockedCodeMapPreRenderService() {
-		codeMapPreRenderService = jest.fn<CodeMapPreRenderService>(() => {
-			return {
-				getRenderMap: jest.fn().mockReturnValue(VALID_NODE_WITH_PATH_AND_EXTENSION),
-				getRenderFileMeta: jest.fn().mockReturnValue(FILE_META)
-			}
-		})()
+		codeMapPreRenderService.getRenderMap = jest.fn().mockReturnValue(VALID_NODE_WITH_PATH_AND_EXTENSION)
+		codeMapPreRenderService.getRenderFileMeta = jest.fn().mockReturnValue(FILE_META)
 	}
 
 	function getFilteredFileContent(name: DownloadCheckboxNames) {