+++ visualization/app/codeCharta/ui/dialog/dialog.service.ts
@@ -1,6 +1,7 @@
 import { dialogDownloadComponent } from "./dialog.download.component"
 import { dialogGlobalSettingsComponent } from "./dialog.globalSettings.component"
 import { addScenarioSettingsComponent } from "./dialog.addScenarioSettings.component"
+import { CCValidationResult } from "../../util/fileValidator"
 
 export class DialogService {
 	/* @ngInject */
@@ -22,19 +23,33 @@ export class DialogService {
 		this.$mdDialog.show(dialog)
 	}
 
-	public showErrorDialog(msg: string = "An error occurred.", title: string = "Error", button: string = "Ok") {
+	public showErrorDialog(msg = "An error occurred.", title = "Error", button = "Ok") {
 		this.$mdDialog.show(this.$mdDialog.alert().clickOutsideToClose(true).title(title).htmlContent(msg).ok(button))
 	}
 
-	public showPromptDialog(
-		msg: string,
-		initial: string,
-		placeholder: string = initial,
-		title: string = "Prompt",
-		button: string = "Ok"
-	): Promise<any> {
+	public showPromptDialog(msg: string, initial: string, placeholder: string = initial, title = "Prompt", button = "Ok"): Promise<any> {
 		const prompt = this.$mdDialog.prompt().title(title).textContent(msg).initialValue(initial).placeholder(placeholder).ok(button)
 
 		return this.$mdDialog.show(prompt)
+	}
+
+	public showValidationWarningDialog(validationResult: CCValidationResult) {
+		const warningSymbol = '<i class="fa fa-exclamation-triangle"></i> '
+
+		const htmlMessage = this.buildHtmlMessage(warningSymbol, validationResult.warning)
+
+		this.showErrorDialog(htmlMessage, validationResult.title)
+	}
+
+	public showValidationErrorDialog(validationResult: CCValidationResult) {
+		const errorSymbol = '<i class="fa fa-exclamation-circle"></i> '
+
+		const htmlMessage = this.buildHtmlMessage(errorSymbol, validationResult.error)
+
+		this.showErrorDialog(htmlMessage, validationResult.title)
+	}
+
+	private buildHtmlMessage(symbol: string, validationResult: string[]): string {
+		return `<p>${validationResult.map(message => symbol + message).join("<br>")}</p>`
 	}
 }