+++ visualization/app/codeCharta/ui/rangeSlider/rangeSlider.component.ts
@@ -19,6 +19,7 @@ import {
 import { FilesService, FilesSelectionSubscriber } from "../../state/store/files/files.service"
 import { isDeltaState } from "../../model/files/files.helper"
 import { FileState } from "../../model/files/files"
+import { BlacklistService, BlacklistSubscriber } from "../../state/store/fileSettings/blacklist/blacklist.service"
 
 export class RangeSliderController
 	implements
@@ -26,14 +27,15 @@ export class RangeSliderController
 		ColorRangeSubscriber,
 		InvertColorRangeSubscriber,
 		WhiteColorBuildingsSubscriber,
-		FilesSelectionSubscriber {
+		FilesSelectionSubscriber,
+		BlacklistSubscriber {
 	private static DEBOUNCE_TIME = 400
 	private readonly applyDebouncedColorRange: (action: SetColorRangeAction) => void
 
-	private DIGIT_WIDTH: number = 11
-	private MIN_DIGITS: number = 4
-	private MAX_DIGITS: number = 6
-	private FULL_WIDTH_SLIDER: number = 235
+	private DIGIT_WIDTH = 11
+	private MIN_DIGITS = 4
+	private MAX_DIGITS = 6
+	private FULL_WIDTH_SLIDER = 235
 
 	private _viewModel: {
 		colorRangeFrom: number
@@ -50,17 +52,26 @@ export class RangeSliderController
 		private $rootScope: IRootScopeService,
 		private $timeout: ITimeoutService,
 		private storeService: StoreService,
-		private metricService: MetricService
+		private metricService: MetricService,
+		private colorRangeService: ColorRangeService
 	) {
 		ColorMetricService.subscribe(this.$rootScope, this)
 		ColorRangeService.subscribe(this.$rootScope, this)
 		InvertColorRangeService.subscribe(this.$rootScope, this)
 		WhiteColorBuildingsService.subscribe(this.$rootScope, this)
 		FilesService.subscribe(this.$rootScope, this)
+		BlacklistService.subscribe(this.$rootScope, this)
 
 		this.applyDebouncedColorRange = _.debounce((action: SetColorRangeAction) => {
 			this.storeService.dispatch(action)
 		}, RangeSliderController.DEBOUNCE_TIME)
+	}
+
+	onBlacklistChanged() {
+		if (this.isMaxMetricValueChanged()) {
+			this.updateMaxMetricValue()
+			this.colorRangeService.reset()
+		}
 	}
 
 	public onColorMetricChanged(colorMetric: string) {
@@ -111,6 +122,11 @@ export class RangeSliderController
 		this._viewModel.sliderOptions.ceil = this.metricService.getMaxMetricByMetricName(
 			this.storeService.getState().dynamicSettings.colorMetric
 		)
+	}
+
+	private isMaxMetricValueChanged() {
+		const newMaxValue: number = this.metricService.getMaxMetricByMetricName(this.storeService.getState().dynamicSettings.colorMetric)
+		return this._viewModel.sliderOptions.ceil !== newMaxValue
 	}
 
 	private initSliderOptions() {